cors/CORS_007.htm

<!DOCTYPE html>
<html>
    <head>
        <title>Author Request Headers do not appear in a preflight for XHR requests</title>
        <meta http-equiv="content-type" content="text/html; charset=UTF-8" />
        <meta description="Verify Author Request Headers do not appear in a preflight for XHR requests" />
        <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
        <link rel="help" href="http://dvcs.w3.org/hg/cors/raw-file/tip/Overview.html#cross-origin-request-with-preflight-0"/>
        <script src="http://w3c-test.org/resources/testharness.js"></script>
        <script src="http://w3c-test.org/resources/testharnessreport.js"></script>
        <script src="SetCORSUrls.js"></script>
    </head>
    <body>
        <div id=log></div>
        <script type="text/javascript">
            var t = async_test("Author Request Headers do not appear in a preflight for XHR requests");
            var xhr = new XMLHttpRequest();
            if('withCredentials' in xhr)
            {
                var randID = Math.floor(Math.random() * 100000);
                var url = CORS_SERVER_URL + "?TestID=7&RequestID=7"+randID;
                xhr.open("GET", url, true);
                xhr.setRequestHeader('X-TEST','value');
                xhr.onreadystatechange = function(e)
                {
                    if (xhr.readyState == 4)
                    {
                        // Because a non-simple author request header ('X-TEST') is being used, the
                        // user-agent should send out a preflight request.
                        // If the preflight request contains the header 'x-test',
                        // the server responds back with 'fail'. Else it responds back 'pass'
                        t.step(function(){assert_true(xhr.status === 200 && xhr.responseText === "pass")});
                        t.done();
                    }
                }
                xhr.send();
            }
            else
            {
                t.step(function(){assert_unreached("[Cross-Origin XMLHttpRequests are not supported]")});
                t.done();
            }
        </script>
    </body>
</html>
